Copy-Item only if path exists

+++ CopyFiles.ps1
@@ -4,6 +4,17 @@
 # Set a flag to force verbose as a default
 $VerbosePreference ='Continue' # equiv to -verbose
 
-Copy-Item -Path "$env:ChocolateyInstall\log" -Destination "$env:BUILD_ARTIFACTSTAGINGDIRECTORY" –Recurse
-Copy-Item -Path "$env:TEMP\chocolatey" -Destination "$env:BUILD_ARTIFACTSTAGINGDIRECTORY" –Recurse
-Copy-Item -Path "$env:ChocolateyInstall\config" -Destination "$env:BUILD_ARTIFACTSTAGINGDIRECTORY" –Recurse
+if (Test-Path "$env:ChocolateyInstall\log") 
+{
+    Copy-Item -Path "$env:ChocolateyInstall\log" -Destination "$env:BUILD_ARTIFACTSTAGINGDIRECTORY" –Recurse
+}
+
+if (Test-Path "$env:TEMP\chocolatey") 
+{
+    Copy-Item -Path "$env:TEMP\chocolatey" -Destination "$env:BUILD_ARTIFACTSTAGINGDIRECTORY" –Recurse
+}
+
+if (Test-Path "$env:ChocolateyInstall\config") 
+{
+    Copy-Item -Path "$env:ChocolateyInstall\config" -Destination "$env:BUILD_ARTIFACTSTAGINGDIRECTORY" –Recurse
+}
